Minimize the diff
Keep the comments template module in place and import it into the
block settings. Drop the controlled inner blocks exclusion from the
block type template fallback: no block type both controls its inner
blocks and declares a template, and the fallback matches how a
template prop already behaves for controlled inner blocks.
