+++ packages/block-editor/src/components/inner-blocks/index.js
@@ -63,8 +63,6 @@ function UncontrolledInnerBlocks( props ) {
 		__experimentalDirectInsert,
 		template,
 		templateLock,
-		value,
-		onChange,
 		wrapperRef,
 		templateInsertUpdatesSelection,
 		__experimentalCaptureToolbars: captureToolbars,
@@ -98,10 +96,7 @@ function UncontrolledInnerBlocks( props ) {
 		clientId,
 		template,
 		templateLock,
-		templateInsertUpdatesSelection,
-		// Controlled inner blocks receive their content from an entity, so
-		// the block type template does not apply to them.
-		! ( value && onChange )
+		templateInsertUpdatesSelection
 	);
 
 	const defaultLayoutBlockSupport =
@@ -160,9 +155,6 @@ function ControlledInnerBlocks( props ) {
 	useBlockSync( props );
 	return <UncontrolledInnerBlocks { ...props } />;
 }
-
-// Exported for unit tests.
-export { ControlledInnerBlocks as __privateControlledInnerBlocks };
 
 const ForwardedInnerBlocks = forwardRef( ( props, ref ) => {
 	const innerBlocksProps = useInnerBlocksProps( { ref }, props );
